Refactor test package names to match code package names
