+++ src/main/java/tf/gpx/edit/viewer/charts/SlopeChart.java
@@ -62,6 +62,10 @@ public class SlopeChart extends HeightChart {
     }
 
     // here we change the series whenever the SlopeBin (or better the SlopeBinColor) changes
+    // in order to speed things up we
+    // - don't check every waypoint
+    // - interpolate the slope-curve using Douglas-Peucker
+    @Override
     protected List<XYChart.Series<Number, Number>> getXYChartSeriesForGPXLineItem(final GPXLineItem lineItem) {
         final List<XYChart.Series<Number, Number>> result = new ArrayList<>();
 
@@ -75,7 +79,6 @@ public class SlopeChart extends HeightChart {
             // y value is the same as in height chart - the elevation
             final double yValue = getYValueAndSetMinMax(gpxWaypoint);
             
-//            System.out.println("adding chart point: " + getMaximumDistance() / 1000.0 + ", " + yValue);
             XYChart.Data<Number, Number> data = new XYChart.Data<>(maxDistance/ 1000.0, yValue);
             data.setExtraValue(gpxWaypoint);
             
@@ -86,6 +89,7 @@ public class SlopeChart extends HeightChart {
             
             // now check for change in color and add series if required
             if (!actColor.equals(slopeColor)) {
+//                System.out.println("New color: " + ColorConverter.JavaFXtoRGBHex(actColor) + " for slope: " + gpxWaypoint.getSlope());
                 // add the current series to the result set
                 if (hasData) {
                     final XYChart.Series<Number, Number> series = new XYChart.Series<>();
@@ -120,13 +124,19 @@ public class SlopeChart extends HeightChart {
     }
     
     // here we set the color of the series to the SlopeBin color and not the color of the linestyle
+    @Override
     protected void setSeriesUserData(final XYChart.Series<Number, Number> series, final GPXLineItem lineItem) {
         String seriesID = lineItem.getCombinedID();
         // add track id for track segments
         if (lineItem.isGPXTrackSegment()) {
             seriesID = lineItem.getParent().getCombinedID() + "." + seriesID;
         }
-        // TFE, 20250502: switched to hex color to be used in css. Needs to be hex to derive 20% opacity values easily
         series.setName(seriesID + DATA_SEP + ColorConverter.JavaFXtoRGBHex(slopeColor));
+    }
+
+    
+    @Override
+    protected String getAreaFillSuffix() {
+        return "";
     }
 }
