+++ src/test/java/tf/gpx/edit/algorithms/smoother/TestDoubleExponentialSmoother.java
@@ -23,13 +23,12 @@
  *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
  *  THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package tf.gpx.edit.algorithms;
+package tf.gpx.edit.algorithms.smoother;
 
 import java.util.Arrays;
 import java.util.List;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import tf.gpx.edit.algorithms.smoother.DoubleExponentialSmoother;
 
 /**
  * Test of the single exponential smoothing implementation.
